Fix text input focussing on keypress
Previous implementation fired the key event on the input successfully,
so the letter you pressed made it into the input - but it didn't then
focus the input, so you couldn't keep typing.

It also only checked that the spans containing text didn't exist -
meaning once there's a single character in the input, the code stops
running. Instead, check if the input's focused, and focus after firing
the event.