Add image pasting functionality
Safari doesn't implement DataTransfer.items, so there's no way to
extract image data (i.e. a screenshot copied to the clipboard) in
Javascript. You can get copied files, but not data. Capture the paste
in Swift and then forward on the image data to the webview.

Currently supports pasting image data (and forwards on text), but
pasting an image file doesn't work.